work

+++ libft/libft.h
@@ -20,7 +20,7 @@ int		ft_isdigit(int x);
 int		ft_isalnum(int x);
 int		ft_isascii(int i);
 int		ft_isprint(int i);
-int		ft_strlen(const char *i);
+size_t	ft_strlen(const char *i);
 void	*ft_memset(void *s, int c, size_t n);
 void	ft_bzero(void *s, size_t n);
 void	*ft_memcpy(void *dest, const void *src, size_t n);
